requirements.txt
numpy==1.26.3
scikit-learn==1.3.2
streamlit==1.29.0
streamlit-option-menu==0.3.6
google-generativeai
python-dotenv
bcrypt==4.1.2